[ORGA] Move students components in dedicated folder
