+++ orga/app/templates/authenticated/sup-students/list.hbs
@@ -1,12 +1,12 @@
 {{page-title (t 'pages.students-sup.title')}}
 <div class="list-students-page">
-  <Routes::Authenticated::SupStudents::ListItems
+  <Student::Sup::List
     @students={{@model}}
     @isLoading={{this.isLoading}}
-    @importStudents={{this.importStudents}}
-    @triggerFiltering={{this.triggerFiltering}}
     @lastNameFilter={{this.lastName}}
     @firstNameFilter={{this.firstName}}
     @studentNumberFilter={{this.studentNumber}}
+    @onImportStudents={{this.importStudents}}
+    @onFilter={{this.triggerFiltering}}
   />
 </div>
